requirements.txt
snowflake-connector-python==3.6.0
requests
apache-airflow-providers-snowflake==5.4.0
apache-airflow-providers-http
apache-airflow-providers-github
pandas
numpy